% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_paper_md.R
\name{make_paper_md}
\alias{make_paper_md}
\title{Make markdown documents for publications}
\usage{
make_paper_md(data, dir = "./content/publication", create = TRUE)
}
\arguments{
\item{data}{The URL or a Sheet ID, which ident}

\item{dir}{The directory you would like the markdown documents to be saved
in. Defaults to the correct folder for publication files when using the
Hugo Academic blogdown template.}

\item{create}{Logicial. If the specified directory doesn't exists, should it
be created? Default is TRUE.}
}
\description{
This function will create markdown documents in the Hugo Academic style for papers
specified in a Google Sheet.
}
