% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_paper_bib.R
\name{make_paper_bib}
\alias{make_paper_bib}
\title{Make .bib files for papers}
\usage{
make_paper_bib(dir, id = "1HPQDH3tOXtZb1DV--8wR9CKAzUz5aywWc2vM3OQ5SrU")
}
\arguments{
\item{dir}{The directory you would like the markdown documents to be saved in}

\item{id}{Your Google Sheet id}
}
\description{
This function will create .bib files in the Hugo Academic style for papers
specified in a Google Sheet.
}
