% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_talk_md.R
\name{make_talk_md}
\alias{make_talk_md}
\title{Make markdown files for talks}
\usage{
make_talk_md(dir, id = "1-PItelqpv0Sb_LdiEDqb8O3D_Roii5nVTL07IRVbRtA")
}
\arguments{
\item{dir}{The directory you would like the markdown documents to be saved in}

\item{id}{Your Google Sheet id}
}
\description{
This function will create markdown files in the Hugo Academic style for talks
specified in a Google Sheet.
}
